services/bridge/requirements.txt
paho-mqtt
sqlalchemy
psycopg2-binary
python-dotenv
